feat(apis/websocket): clear old client sessions and instances
