+++ packages/shared/src/constants/HumanizedDisconnectReason.ts
@@ -12,6 +12,7 @@ const HumanizedDisconnectReason = {
     [DisconnectReason.TooSlow]: 'the client was not ready in time',
     [DisconnectReason.PlannedDisconnect]: 'the client has disconnected on its own',
     [DisconnectReason.NoOpenSocket]: 'the receiving end did not have an open socket',
+    [DisconnectReason.NewConnection]: 'the client connected from another location',
 } as const satisfies Record<DisconnectReason | number, string>
 
 export default HumanizedDisconnectReason
